requirements copy.txt
asgiref==3.5.2
backcall==0.2.0
backports.zoneinfo==0.2.1
click==7.1.2
decorator==5.0.7
Django==4.1.1
dlib==19.22.99
face-recognition==1.3.0
face-recognition-models==0.3.0
imutils==0.5.4
ipdb==0.13.7
ipython==7.23.1
ipython-genutils==0.2.0
jedi==0.18.0
matplotlib-inline==0.1.2
MouseInfo==0.1.3
numpy==1.20.2
opencv-python==4.5.2.52
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
prompt-toolkit==3.0.18
ptyprocess==0.7.0
PyAutoGUI==0.9.52
PyGetWindow==0.0.9
Pygments==2.9.0
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.1.4
PyScreeze==0.1.27
python3-xlib==0.15
PyTweening==1.0.3
PyYAML==5.4.1
sqlparse==0.4.2
toml==0.10.2
traitlets==5.0.5
wcwidth==0.2.5